+++ fixtures/LIN3S/WPFoundation/Acf.php
@@ -1,0 +1,32 @@
+<?php
+
+/*
+ * This file is part of the WPFoundation library.
+ *
+ * Copyright (c) 2015-2016 LIN3S <info@lin3s.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace fixtures\LIN3S\WPFoundation;
+
+use LIN3S\WPFoundation\Configuration\Acf\Acf as BaseAcf;
+
+/**
+ * Dummy implementation of ACF class for fixture purposes.
+ *
+ * @author Beñat Espiña <benatespina@gmail.com>
+ */
+class Acf extends BaseAcf
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function wyswygToolbars()
+    {
+        return [
+            'lin3s' => [1 => ['bold', 'italic', 'bullist', 'numlist', 'link', 'unlink']],
+        ];
+    }
+}